+++ tests/agent_server/test_deploy_trigger.py
@@ -20,76 +20,78 @@ import logging
 import pytest
 
 from agent_server.conftest import _deploy_resources, get_agent, get_resource
-from utils import AsyncClosing, log_contains, log_doesnt_contain, retry_limited
+from utils import log_contains, log_doesnt_contain, retry_limited
 
 
 @pytest.mark.asyncio(timeout=150)
-async def test_deploy_trigger(server, client, resource_container, environment, caplog, no_agent_backoff):
+async def test_deploy_trigger(
+    server, client, clienthelper, resource_container, environment, caplog, no_agent_backoff, async_finalizer
+):
     """
        Test deployment of empty model
     """
     caplog.set_level(logging.INFO)
-    async with AsyncClosing(get_agent(server, environment, "agent1", "agent5")):
 
-        version = 5
+    agent = await get_agent(server, environment, "agent1", "agent5")
+    async_finalizer(agent.stop)
 
-        resources = [
-            get_resource(version, agent="agent1"),
-            get_resource(version, agent="agent2"),
-            get_resource(version, agent="agent3"),
-        ]
+    version = await clienthelper.get_version()
 
-        await _deploy_resources(client, environment, resources, version, False)
+    resources = [
+        get_resource(version, agent="agent1"),
+        get_resource(version, agent="agent2"),
+        get_resource(version, agent="agent3"),
+    ]
 
-        async def verify(result, a1=0, code=200, warnings=["Could not reach agents named [agent2,agent3]"], agents=["agent1"]):
-            assert result.code == code
+    await _deploy_resources(client, environment, resources, version, False)
 
-            def is_deployed():
-                return resource_container.Provider.readcount("agent1", "key1") == a1
+    async def verify(result, a1=0, code=200, warnings=["Could not reach agents named [agent2,agent3]"], agents=["agent1"]):
+        assert result.code == code
 
-            await retry_limited(is_deployed, 1)
-            log_contains(caplog, "agent", logging.INFO, f"Agent agent1 got a trigger to update in environment {environment}")
-            log_doesnt_contain(
-                caplog, "agent", logging.INFO, f"Agent agent5 got a trigger to update in environment {environment}"
-            )
+        def is_deployed():
+            return resource_container.Provider.readcount("agent1", "key1") == a1
 
-            assert result.result["agents"] == agents
-            if warnings:
-                assert sorted(result.result["metadata"]["warnings"]) == sorted(warnings)
-            caplog.clear()
+        await retry_limited(is_deployed, 1)
+        log_contains(caplog, "agent", logging.INFO, f"Agent agent1 got a trigger to update in environment {environment}")
+        log_doesnt_contain(caplog, "agent", logging.INFO, f"Agent agent5 got a trigger to update in environment {environment}")
 
-        async def verify_failed(result, code=400, message="", warnings=["Could not reach agents named [agent2,agent3]"]):
-            assert result.code == code
+        assert result.result["agents"] == agents
+        if warnings:
+            assert sorted(result.result["metadata"]["warnings"]) == sorted(warnings)
+        caplog.clear()
 
-            log_doesnt_contain(caplog, "agent", logging.INFO, "got a trigger to update")
-            if warnings:
-                assert sorted(result.result["metadata"]["warnings"]) == sorted(warnings)
-            assert result.result["message"] == message
-            caplog.clear()
+    async def verify_failed(result, code=400, message="", warnings=["Could not reach agents named [agent2,agent3]"]):
+        assert result.code == code
 
-        # normal
-        result = await client.deploy(environment)
-        await verify(result, a1=1)
+        log_doesnt_contain(caplog, "agent", logging.INFO, "got a trigger to update")
+        if warnings:
+            assert sorted(result.result["metadata"]["warnings"]) == sorted(warnings)
+        assert result.result["message"] == message
+        caplog.clear()
 
-        # only agent1
-        result = await client.deploy(environment, agents=["agent1"])
-        await verify(result, a1=2, warnings=None)
+    # normal
+    result = await client.deploy(environment)
+    await verify(result, a1=1)
 
-        # only agent5 (not in model)
-        result = await client.deploy(environment, agents=["agent5"])
-        await verify_failed(
-            result, 404, "No agent could be reached", warnings=["Model version 5 does not contain agents named [agent5]"]
-        )
+    # only agent1
+    result = await client.deploy(environment, agents=["agent1"])
+    await verify(result, a1=2, warnings=None)
 
-        # only agent2 (not alive)
-        result = await client.deploy(environment, agents=["agent2"])
-        await verify_failed(result, 404, "No agent could be reached", warnings=["Could not reach agents named [agent2]"])
+    # only agent5 (not in model)
+    result = await client.deploy(environment, agents=["agent5"])
+    await verify_failed(
+        result, 404, "No agent could be reached", warnings=[f"Model version {version} does not contain agents named [agent5]"]
+    )
 
-        # All of it
-        result = await client.deploy(environment, agents=["agent1", "agent2", "agent5"])
-        await verify(
-            result,
-            a1=3,
-            agents=["agent1"],
-            warnings=["Could not reach agents named [agent2]", "Model version 5 does not contain agents named [agent5]"],
-        )
+    # only agent2 (not alive)
+    result = await client.deploy(environment, agents=["agent2"])
+    await verify_failed(result, 404, "No agent could be reached", warnings=["Could not reach agents named [agent2]"])
+
+    # All of it
+    result = await client.deploy(environment, agents=["agent1", "agent2", "agent5"])
+    await verify(
+        result,
+        a1=3,
+        agents=["agent1"],
+        warnings=["Could not reach agents named [agent2]", f"Model version {version} does not contain agents named [agent5]"],
+    )