+++ tests/agent_server/test_send_events.py
@@ -17,7 +17,6 @@
 """
 import asyncio
 import logging
-import time
 
 import pytest
 
@@ -25,19 +24,20 @@ from agent_server.conftest import _wait_for_n_deploying
 from inmanta import config, const
 from inmanta.agent.agent import Agent
 from inmanta.server import SLICE_AGENT_MANAGER
+from inmanta.util import get_compiler_version
 from utils import _wait_until_deployment_finishes, retry_limited
 
 logger = logging.getLogger("inmanta.test.send_events")
 
 
 @pytest.mark.asyncio
-async def test_send_events(resource_container, environment, server, client, agent):
+async def test_send_events(resource_container, environment, server, client, agent, clienthelper):
     """
         Send and receive events within one agent
     """
     resource_container.Provider.reset()
 
-    version = int(time.time())
+    version = await clienthelper.get_version()
 
     res_id_1 = "test::Resource[agent1,key=key1],v=%d" % version
     resources = [
@@ -59,7 +59,14 @@ async def test_send_events(resource_container, environment, server, client, agen
         },
     ]
 
-    result = await client.put_version(tid=environment, version=version, resources=resources, unknowns=[], version_info={})
+    result = await client.put_version(
+        tid=environment,
+        version=version,
+        resources=resources,
+        unknowns=[],
+        version_info={},
+        compiler_version=get_compiler_version(),
+    )
     assert result.code == 200
 
     # do a deploy
@@ -81,7 +88,7 @@ async def test_send_events(resource_container, environment, server, client, agen
 
 
 @pytest.mark.asyncio
-async def test_send_events_cross_agent(resource_container, environment, server, client, async_finalizer):
+async def test_send_events_cross_agent(resource_container, environment, server, client, async_finalizer, clienthelper):
     """
         Send and receive events over agents
     """
@@ -100,7 +107,7 @@ async def test_send_events_cross_agent(resource_container, environment, server, 
     await agent2.start()
     await retry_limited(lambda: len(agentmanager.sessions) == 2, 10)
 
-    version = int(time.time())
+    version = await clienthelper.get_version()
 
     res_id_1 = "test::Resource[agent1,key=key1],v=%d" % version
     resources = [
@@ -122,7 +129,14 @@ async def test_send_events_cross_agent(resource_container, environment, server, 
         },
     ]
 
-    result = await client.put_version(tid=environment, version=version, resources=resources, unknowns=[], version_info={})
+    result = await client.put_version(
+        tid=environment,
+        version=version,
+        resources=resources,
+        unknowns=[],
+        version_info={},
+        compiler_version=get_compiler_version(),
+    )
     assert result.code == 200
 
     # do a deploy
@@ -148,7 +162,7 @@ async def test_send_events_cross_agent(resource_container, environment, server, 
 
 @pytest.mark.asyncio
 async def test_send_events_cross_agent_deploying(
-    resource_container, environment, server, client, no_agent_backoff, async_finalizer
+    resource_container, environment, server, client, no_agent_backoff, async_finalizer, clienthelper
 ):
     """
         Send and receive events over agents
@@ -168,7 +182,7 @@ async def test_send_events_cross_agent_deploying(
     await agent2.start()
     await retry_limited(lambda: len(agentmanager.sessions) == 2, 10)
 
-    version = int(time.time())
+    version = await clienthelper.get_version()
 
     res_id_1 = "test::Resource[agent1,key=key1],v=%d" % version
     resources = [
@@ -190,7 +204,14 @@ async def test_send_events_cross_agent_deploying(
         },
     ]
 
-    result = await client.put_version(tid=environment, version=version, resources=resources, unknowns=[], version_info={})
+    result = await client.put_version(
+        tid=environment,
+        version=version,
+        resources=resources,
+        unknowns=[],
+        version_info={},
+        compiler_version=get_compiler_version(),
+    )
     assert result.code == 200
 
     # do a deploy
@@ -214,7 +235,7 @@ async def test_send_events_cross_agent_deploying(
 
 @pytest.mark.asyncio(timeout=15)
 async def test_send_events_cross_agent_restart(
-    resource_container, environment, server, client, no_agent_backoff, async_finalizer
+    resource_container, environment, server, client, clienthelper, no_agent_backoff, async_finalizer
 ):
     """
         Send and receive events over agents with agents starting after deploy
@@ -223,14 +244,16 @@ async def test_send_events_cross_agent_restart(
 
     config.Config.set("config", "agent-deploy-interval", "0")
     config.Config.set("config", "agent-repair-interval", "0")
+
     resource_container.Provider.reset()
+
     agent2 = Agent(hostname="node2", environment=environment, agent_map={"agent2": "localhost"}, code_loader=False)
     async_finalizer.add(agent2.stop)
     agent2.add_end_point_name("agent2")
     await agent2.start()
     await retry_limited(lambda: len(agentmanager.sessions) == 1, 10)
 
-    version = int(time.time())
+    version = await clienthelper.get_version()
 
     res_id_1 = "test::Resource[agent1,key=key1],v=%d" % version
     resources = [
@@ -252,8 +275,7 @@ async def test_send_events_cross_agent_restart(
         },
     ]
 
-    result = await client.put_version(tid=environment, version=version, resources=resources, unknowns=[], version_info={})
-    assert result.code == 200
+    await clienthelper.put_version_simple(resources, version)
 
     # do a deploy
     result = await client.release_version(environment, version, True, const.AgentTriggerMethod.push_full_deploy)